+++ src/bifidotyper/__init__.py
@@ -1,7 +1,19 @@
 from .cli import main
 from .processor import build_sample_dict
 from .references import ReferenceManager
+from .sylph import SylphUtils
+from .hmo_genes import HMOUtils
+from .plotting import PlotUtils
+from .logger import logger
 
-__version__ = "0.1.4"
+__version__ = "0.2.1"
 
-__all__ = ["main", "build_sample_dict", "ReferenceManager","SylphUtils","HMOUtils","PlotUtils","logger"]
+__all__ = [
+	"main",
+	"build_sample_dict",
+	"ReferenceManager",
+	"SylphUtils",
+	"HMOUtils",
+	"PlotUtils",
+	"logger",
+]